项目实现了查询购物车、继续购物、显示购买物品的总价、退出等功能

+++ src/ReadProductExcel.java
@@ -27,7 +27,7 @@ public class ReadProductExcel {
                     } else if (k == 1) {
                         product.setName(this.getValue(cell));
                     } else if (k == 2) {
-                        product.setPrice(this.getValue(cell));
+                        product.setPrice(Float.valueOf(this.getValue(cell)));
                     } else if (k == 3) {
                         product.setDesc(this.getValue(cell));
                     }
@@ -57,7 +57,7 @@ public class ReadProductExcel {
                     } else if (k == 1) {
                         product.setName(this.getValue(cell));
                     } else if (k == 2) {
-                        product.setPrice(this.getValue(cell));
+                        product.setPrice(Float.valueOf(this.getValue(cell)));
                     } else if (k == 3) {
                         product.setDesc(this.getValue(cell));
                     }
